android: move the SDL subclasses into SDL's package
