+++ android/app/src/main/java/org/vita3k/emulator/Emulator.java
@@ -52,6 +52,7 @@ import java.util.Arrays;
 
 import org.libsdl.app.SDLActivity;
 import org.libsdl.app.SDLDummyEdit;
+import org.libsdl.app.VitaTextEdit;
 import org.libsdl.app.SDLSurface;
 import org.vita3k.emulator.data.AppStorage;
 import org.vita3k.emulator.data.NativeImeState;
@@ -765,7 +766,7 @@ public class Emulator extends SDLActivity
 
         mLayout.removeView(mTextEdit);
         final VitaTextEdit replacement = new VitaTextEdit(this);
-        replacement.setInputType(lastTextInputType);
+        replacement.applyInputType(lastTextInputType);
         mTextEdit = replacement;
         mLayout.addView(mTextEdit, params);
 
@@ -798,13 +799,15 @@ public class Emulator extends SDLActivity
 
         if (mTextEdit == null) {
             final VitaTextEdit replacement = new VitaTextEdit(this);
-            replacement.setInputType(lastTextInputType);
+            replacement.applyInputType(lastTextInputType);
             mTextEdit = replacement;
             mLayout.addView(mTextEdit, new RelativeLayout.LayoutParams(lastTextInputLayoutParams));
         } else {
             ensureVitaTextEdit();
             mTextEdit.setLayoutParams(new RelativeLayout.LayoutParams(lastTextInputLayoutParams));
-            mTextEdit.setInputType(lastTextInputType);
+            if (mTextEdit instanceof VitaTextEdit) {
+                ((VitaTextEdit) mTextEdit).applyInputType(lastTextInputType);
+            }
         }
 
         mTextEdit.setVisibility(View.VISIBLE);
